requirements.txt
pygame>=2.1.0
Pillow>=9.0.0
PyYAML>=6.0
Flask>=2.3.0
msal>=1.20.0
requests>=2.28.0
piexif>=1.1.3

# Uncomment on Raspberry Pi for e-ink display support:
# inky[rpi,fonts]>=1.3.0
# RPi.GPIO>=0.7.0
